requirements.txt
torch
torchvision
timm
numpy
opencv-python
nibabel
streamlit
Pillow
matplotlib